Add camp and demo files Move camp code into camp file Move demo variables into demo file
